# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Food Facts API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/status": {
            "get": {
                "tags": [
                    "/status"
                ],
                "summary": "Display a listing of the resource",
                "description": "Get all status on database",
                "operationId": "5e01603cffcb85640997089cfec54763",
                "responses": {
                    "200": {
                        "description": "List of status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produtos": {
            "get": {
                "tags": [
                    "/produtos"
                ],
                "summary": "Display a listing of the resource and paginate then",
                "description": "Get all products on database",
                "operationId": "06ec4c54910f9c241e8c22e51f6e3042",
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produtos/{id}": {
            "get": {
                "tags": [
                    "/produtos"
                ],
                "summary": "Display only one product",
                "description": "Get a product from the database",
                "operationId": "475cb0b7b065fb84b474c06cbdf13a78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show a product"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/produtos"
                ],
                "summary": "Update a product",
                "description": "Update a product on database",
                "operationId": "3b1d0153f7f031da88cc6e3e2f690e61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    },
                                    "brands": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "string"
                                    },
                                    "labels": {
                                        "type": "string"
                                    },
                                    "cities": {
                                        "type": "string"
                                    },
                                    "purchase_places": {
                                        "type": "string"
                                    },
                                    "stores": {
                                        "type": "string"
                                    },
                                    "ingredients_text": {
                                        "type": "string"
                                    },
                                    "traces": {
                                        "type": "string"
                                    },
                                    "serving_size": {
                                        "type": "string"
                                    },
                                    "serving_quantity": {
                                        "type": "integer"
                                    },
                                    "nutriscore_score": {
                                        "type": "integer"
                                    },
                                    "nutriscore_grade": {
                                        "type": "string"
                                    },
                                    "main_category": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a product"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/produtos"
                ],
                "summary": "Change the product status",
                "description": "Change the product status to 'trash'",
                "operationId": "7f151ffdaca32da6bbe069b43680be74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Change the product status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}